requirements.txt
# ----------------------------
# Core Frameworks
# ----------------------------
streamlit
flask

# ----------------------------
# Data Handling & Processing
# ----------------------------
pandas
numpy
openpyxl
XlsxWriter
scikit-learn

# ----------------------------
# Visualization Libraries
# ----------------------------
plotly
kaleido
matplotlib
seaborn

# ----------------------------
# AI Integration & APIs
# ----------------------------
requests
tqdm
python-dotenv
ollama

# ----------------------------
# Reporting & Export
# ----------------------------
reportlab

streamlit requests